Move all home logic into hook
